refactor: 내계정 모듈화
