+++ src/pages/MyAccountPage/index.tsx
@@ -1,24 +1,8 @@
-import React, { useEffect } from 'react';
-import { useNavigate } from 'react-router-dom';
-
 import { Button } from '@/components/common/Button';
-import { useAuth } from '@/context/useAuth';
+import { useLogout } from '@/pages/MyAccountPage/useLogout';
 
-const MyAccountPage: React.FC = () => {
-  const { authToken, logout } = useAuth();
-  const navigate = useNavigate();
-
-  useEffect(() => {
-    if (!authToken) {
-      navigate('/login'); // 비로그인 상태: 로그인 페이지로 리디렉션
-    }
-  }, [authToken, navigate]);
-
-  const handleLogout = () => {
-    logout();
-    alert('로그아웃 처리');
-    navigate('/'); // 메인 페이지로 리디렉션
-  };
+const MyAccountPage = () => {
+  const handleLogout = useLogout();
 
   return (
     <div>
